+++ aisec/constants.go
@@ -2,7 +2,7 @@ package aisec
 
 // Version and user agent.
 const (
-	Version   = "0.3.1"
+	Version   = "0.4.0"
 	UserAgent = "PAN-AIRS/" + Version + "-go-sdk"
 )